% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_model_m.R
\name{forest_model_m}
\alias{forest_model_m}
\title{Creates data for forestplot, 99 pct of code from the forestmodel package}
\usage{
forest_model_m(
  model,
  panels = default_forest_panels(model, factor_separate_line = factor_separate_line),
  covariates = NULL,
  exponentiate = NULL,
  funcs = NULL,
  factor_separate_line = FALSE,
  format_options = forest_model_format_options(),
  theme = theme_forest(),
  limits = NULL,
  breaks = NULL,
  return_data = FALSE,
  recalculate_width = TRUE,
  recalculate_height = TRUE,
  model_list = NULL,
  merge_models = FALSE,
  exclude_infinite_cis = TRUE
)
}
\arguments{
\item{model}{model}
}
\value{
data for plot
}
\description{
This is used to make data for plots
}
\details{
changed some things in the broom coef where it won't correctly give values
}
\examples{
# something

#forest_model_m(model = fit_mul, return_data = T, factor_separate_line = T, exponentiate = T)

}
