% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mforestmodel.R
\name{mforestmodel}
\alias{mforestmodel}
\title{Creates multivariate forestplot}
\usage{
mforestmodel(
  data,
  pala = "grey25",
  palb = "grey75",
  lim = c(-2.6, 4),
  dependent = "delta",
  legend_position = "none",
  header = NULL,
  spaces = NULL,
  est_family = "poisson"
)
}
\arguments{
\item{data}{data to use models}
}
\value{
data for plot
}
\description{
This is used to make data for plots
}
\details{
changed some things in the broom coef where it won't correctly give values
}
\examples{
# something


}
